fix racket dev environments on NixOS
I've had some issues with my dev environments that I think comes from updates to packages in my nix-shell environment for racket development.  Perhaps the root of this is that Racket will stash some path into /nix/store/ and then when a package in the nix-shell is updated that path is no longer valid.

I think this problem wouldn't exist if libraries could just be linked into /run/current-system, but no!  Anyway, I'm now using a pinned version of nixpkgs so that the build dependencies can change under me.  So now my environment is *really* reproducible.  And so much more of a hassle than if libraries specified in systemPackages were simply linked into the current system environment!