racket dev: fix nixos environment to something that will maybe work
Switching to closer-to-normal nix-shell instead of FHSUserEnv or
trying to use the system packages.  But still trying to use system
packages for emacs, etc.  If this goes poorly I will switch to all
tools being pinned.

However, I still need to futz with the nix-shell environment to get
LD_LIBRARY_PATH to include libcrypto.so, and to NOT disable CA-certs
in the environment (needed for raco to resolve packages).