add tmux-red config and commands to enable nested tmux (one level)
It's not a very principled way of doing this.  I've been meaning to do
something like this for a while and will probably do it better
eventually, but out of expedience I need to nest two levels of tmux
imminently.

I should decide on a few prefix keys and colors to go with them to
have nestability up to several levels, and write more generic code to
support it.