emacs: add xml handiness
Since I'm thinking about tree editing, I'm looking at what's available
for XML.  This was the lowest-hanging fruit that will be nice if/when
I ever need to seriously edit XML... which maybe I won't very often,
so... maybe I shouldn't waste time on this.  But in doing this tree
operation stuff I want to have more example trees to think about and
operate with.