emacs: add more submodules
I may as well have submodules for not only the very core packages that I use, but the secondary ones as well.  I will not add the tertiary packages that I use occasionally or that are mode-specific, though I am adding systemd mode, because I'll likely want to use that on servers where I only use these submodule packages.