add hyprland-fix-lock-crash
Sigh.  Every Wayland window manager feels like a step back from Awesomewm and X11 in some respects, but also Wayland improves some things and Hyprland does a couple of things that I like better than Awesomewm.  But it is so much less stable in other ways.  I'm sick of the frequency of the lock screen crashing.